Version 1.0
Pumped up and ready to kill!

The new version of owrpc is here, a much tidier and more stable improvement from maxicc/overwatch-rpc! Currently you can set a Discord presence for Quick Play, Competitive or Arcade. You can also set when you're In Menus or In Queue, and you can also make a custom presence, for example if you're in skrims or a custom game!